refactor(mall-core): 重构订单相关代码
- 修改了多个类中的字段类型，如将 Double 改为 BigDecimal，Integer 改为 Long
- 优化了订单总价的计算逻辑，使用 BigDecimal进行精确计算
- 新增了 DecimalUtils 工具类，用于 BigDecimal 的加法和除法运算
- 删除了未使用的 Order 配置类
- 优化了支付宝和微信支付的逻辑

+++ mall-core/src/main/java/com/jiang/mall/dao/CartMapper.java
@@ -37,7 +37,7 @@ public interface CartMapper extends BaseMapper<Cart> {
 	Cart selectOneByProdIdAndUserId(Long productId, Long userId);
 
 	@Update("UPDATE tb_carts SET num = num + #{num} WHERE id = #{id}")
-	int updateNumById(Long id, Integer num);
+	int updateNumById(Long id, Long num);
 
 	@Select("SELECT user_id FROM tb_carts WHERE id = #{id} LIMIT 1")
 	Long selectUserIdById(Long id);
