+++ mall-core/src/main/java/com/jiang/mall/domain/vo/CheckoutVo.java
@@ -43,7 +43,7 @@ public class CheckoutVo {
     /**
      * 商品数量
      */
-    private Integer num;
+    private Long num;
 
 }
