+++ mall-core/src/main/java/com/jiang/mall/service/ICartService.java
@@ -76,7 +76,7 @@ public interface ICartService extends IService<Cart> {
      * @param sessionId 用户会话ID
      * @return 布尔值，表示购物车记录是否成功插入或更新
      */
-	Boolean insertCart(Long productId, Integer num, String sessionId);
+	Boolean insertCart(Long productId, Long num, String sessionId);
 
 	/**
      * 更新购物车中商品的数量
@@ -89,7 +89,7 @@ public interface ICartService extends IService<Cart> {
      * @param sessionId 用户的会话ID，用于识别和验证用户
      * @return 如果商品不属于当前用户，返回null；否则，返回更新是否成功的布尔值
      */
-	Boolean updateCart(Long id, Integer num, String sessionId);
+	Boolean updateCart(Long id, Long num, String sessionId);
 
 	/**
      * 删除购物车项
